--- pages/failbg/index.json
@@ -1,5 +1,0 @@
-{
-  "navigationBarTitleText": "失败惹",
-  "usingComponents": {
-  }
-}
