+++ pages/index/index.json
@@ -1,16 +1,10 @@
 {
-  "navigationBarTitleText": "AI屋GI物",
+  "navigationBarTitleText": "物品识别",
   "usingComponents": {
     "wux-row": "../../miniprogram_npm/wux-weapp/row/index",
     "wux-col": "../../miniprogram_npm/wux-weapp/col/index",
-    "wux-popup": "../../miniprogram_npm/wux-weapp/popup/index",
-    "wux-divider": "../../miniprogram_npm/wux-weapp/divider/index",
-    "wux-button": "../../miniprogram_npm/wux-weapp/button/index",
     "wux-wing-blank": "../../miniprogram_npm/wux-weapp/wing-blank/index",
     "wux-white-space": "../../miniprogram_npm/wux-weapp/white-space/index",
-    "wux-grids": "../../miniprogram_npm/wux-weapp/grids/index",
-    "wux-grid": "../../miniprogram_npm/wux-weapp/grid/index",
-    "wux-cell-group": "../../miniprogram_npm/wux-weapp/cell-group/index",
-    "wux-cell": "../../miniprogram_npm/wux-weapp/cell/index"
+    "wux-progress": "../../miniprogram_npm/wux-weapp/progress/index"
   }
 }
