--- pages/mobilenet/index.wxml
@@ -1,23 +1,0 @@
-<view class="page">
-  <view class="page__bd">
-    <camera
-      style="display:{{camera?'block':'none'}}"
-      device-position="back"
-      flash="off"
-      binderror="error"
-      frame-size="small"
-      class="camera">
-    </camera>
-    <wiew class="result">
-      <view class="text">{{result?'您成功识别了':'预测结果'}}：{{result || resultOb.label}}</view>
-      <view class="progress">
-        <wux-progress  active-color="#e96900" status="progress" percent="{{ resultOb.value * 100 }}" />
-      </view>
-    </wiew>
-    
-    <!-- <wiew>运算次数：{{cacular}}</wiew> -->
-    <view wx:if="{{!camera}}" class="graph_wall">
-      <image class="img" src="{{src}}" alt="photo"></image>
-    </view>
-  </view>
-</view>
