--- pages/mobilenet/index.wxss
@@ -1,100 +1,0 @@
-/* pages/mobilenet/index.wxss */
-page{
-  width: 100%;
-  height: 100%;
-  background-color: rgba(255,255,255,1);
-}
-
-.camera {
-  width: 75%;
-  height: 100vw;
-  background-color: #fff;
-  margin: 20% auto;
-  z-index: 5;
-}
-
-.prediction-result {
-  height: calc(100vh - 100vw); 
-}
-.cover-view{
-  width: 100%;
-  height: 92vh;
-  font-size: 0;
-  position: relative;
-}
-.mask{
-  float: left;
-  width: 20%;
-  height: 92vh;
-  background: rgba(0,0,0,.4);
-}
-.mid-mask{
-  float: left;
-  width: 60%;
-  height: 92vh;
-}
-.mid-mask-top-bottom{
-  width: 100%;
-  height: 6vh;
-  background: rgba(0,0,0,.4);
-}
-.scan-aera{
-  width: 100%;
-  height: 80vh;
-  background: rgba(255, 255, 255, 0);
-}
-
-.logs{
-  position: absolute;
-  top: 15px;
-  left: 10px;
-  color: brown;
-}
-.graph_wall{
-  width: 75%;
-  height: 100vw;
-  margin: 20% auto;
-}
-.graph_wall .img {
-  width: 100%;
-  height: 100%;
-  z-index: 5;
-}
-.graph_wall .mask{
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100vw;
-  height: 100vh;
-  z-index: 10;
-  background: rgba(0,0,0,.4);
-}
-.graph_wall .mask .content{
-  position: absolute;
-  bottom: 20%;
-  left: 50%;
-  z-index: 20;
-  width: 70vw;
-  text-align: center;
-  color: #333;
-  transform: translateX(-50%);
-  background: rgba(255, 255, 255, 0);
-  font-size: 24px;
-}
-.result{
-  font-weight: bolder;
-  display:block;
-  width:100%;
-  text-align:center;
-  color:#333;
-  position: absolute;
-  bottom: 10vh;
-  left: 0;
-  width: 100%;
-  font-size: 18px;
-  letter-spacing: 1px;
-}
-.progress{
-  margin: 5px auto;
-  width: 70vw;
-}
