--- pages/video/index.json
@@ -1,6 +1,0 @@
-{
-  "navigationBarTitleText": "视频",
-  "usingComponents": {
-
-  }
-}
