--- pages/video/index.wxss
@@ -1,40 +1,0 @@
-#myVideo{
-  width: 100%;
-  height: 100vh;
-}
-.mask{
-  width: 100%;
-  height: 100vh;
-  position: absolute;
-  bottom: 0px;
-  text-align: center;
-  
-}
-.description{
-  font-size: 22px;
-  color: #FFF;
-  white-space: normal;
-  overflow: visible;
-  line-height: 15vh;
-}
-.maskBg{
-  position: absolute;
-  bottom: 0px;
-  height: 15vh;
-  background: rgba(33,33,33,.4);
-  width: 100%;
-}
-.btnBack{
-  width: 186px;
-  padding: 10px;
-  color: #FFF;
-  background: rgba(0,0,0,.8);
-  text-align: center;
-  border-radius: 5px;
-  position: absolute;
-  top: 45%;
-  left: 50%;
-  transform: translateX(-93px);
-  letter-spacing: 1px;
-  z-index: 9999;
-}
